rag-chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2

streamlit
requests

langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb

pypdf